+++ tests/unit/data/image/test_mvtec_3d.py
@@ -31,3 +31,8 @@ class TestMVTec3D(_TestAnomalibDepthDatamodule):
         _datamodule.setup()
 
         return _datamodule
+
+    @pytest.fixture()
+    def fxt_data_config_path(self) -> str:
+        """Return the path to the test data config."""
+        return "configs/data/mvtec_3d.yaml"